+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes
 
+## v1.0.3
+
+Release date: TBD
+
+* Dropped support of end of life MySQL version 5.6
+* Dropped support of end of life MariaDB versions below 10.3
+* Dropped support of end of life Python version 3.6
+
+
 ## v1.0.2
 
 Release date: 2021-01-09
@@ -195,7 +204,7 @@ Release date: 2016-08-30
 Release date: 2016-07-29
 
 * Fix SELECT JSON type cause UnicodeError
-* Avoid float convertion while parsing microseconds
+* Avoid float conversion while parsing microseconds
 * Warning has number
 * SSCursor supports warnings